Limit config var limits like pages fetches
Using the "limit" config variable, the number of like pages fetched is
limited. Note that like pages are in batches of 50, so a limit of 20 (now
the default value) will limit the number of page requests to 1.

While testing this I received an error in the form of cloudflare blocking
one (or more) of my requests. I waited a few minutes and haven't accounted
it again yet. May need to limit the number of concurrent requests, easy to
do with Async, may add to config variables.