Functional Electron UI
Very basic UI, built on top of Electron.

Renderer process calls main process to download images once user has set
config values as desired.

Next up is to add progress dialog and loading dialogs for when waiting on
events between renderer and main processes to come through. After that is
probably just the packaged & rebranded Electron executable, and throwing
it in an installer, probably use NSIS or something for that. Also need to
save config on exit.